feat : parcours_orientation v1 + liste de parcours

+++ backend/api/enfiletesbaskets/src/main/java/com/enfiletesbaskets/enfiletesbaskets/repositories/ActualityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActualityRepository extends JpaRepository<ActualityModel, Long> {
+public interface ActualityRepository extends JpaRepository<ActualityModel, Integer> {
     // Ajoutez des méthodes personnalisées si nécessaire
 }
