+++ backend/api/enfiletesbaskets/src/main/java/com/enfiletesbaskets/enfiletesbaskets/repositories/CategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<CategoryModel, Long> {
+public interface CategoryRepository extends JpaRepository<CategoryModel, Integer> {
     // Ajoutez des méthodes personnalisées si nécessaire
 }
