+++ backend/api/enfiletesbaskets/src/main/java/com/enfiletesbaskets/enfiletesbaskets/repositories/CommunityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommunityRepository extends JpaRepository<CommunityModel, Long> {
+public interface CommunityRepository extends JpaRepository<CommunityModel, Integer> {
     // Ajoutez des méthodes personnalisées si nécessaire
 }
